requirements.txt
Pillow>=7.10
python-dateutil>=2.8.0
python-dotenv>=0.13.0
rafraser-theia>=0.0.1
requests